Rename src/bridge/rsa to src/bridge/crypto
