+++ src/bridge/crypto/ec.h
